requirement.txt
nextcord
pillow
requests
python-dotenv
pymongo